feat: implemented cook's dashboard UI

+++ frontend/src/pages/CookDashboard.js
@@ -1,7 +1,13 @@
-import { useState, useEffect } from "react";
+import { useState, useEffect, useContext } from "react";
 import API from "../services/api";
+import { toast } from "react-toastify";
+import { AuthContext } from "../context/AuthContext";
+import logo from "../assets/logo.png";
+import "../styles/cook.css";
 
 function CookDashboard(){
+    const {user} = useContext(AuthContext);
+    const {logout} = useContext(AuthContext);
     const [myFoods, setMyFoods] = useState([]);
     const [editingFood, setEditingFood] = useState(null);
     const [form, setForm] = useState({
@@ -17,10 +23,7 @@ function CookDashboard(){
     });
 
     const [orders, setOrders] = useState([]);
-    const [incomingOrders, setIncomingOrders] = useState([]);
-    const [activeOrders, setActiveOrders] = useState([]);
-    const [completeOrders, setCompleteOrders] = useState([]);
-    const [cancelledOrders, setCancelledOrders] = useState([]); 
+    const [search, setSearch] = useState("");
     const [activeTab, setActiveTab] = useState("incoming");
     //handle form input
     const handleChange = (e)=>{
@@ -34,12 +37,12 @@ function CookDashboard(){
         try{
             if(editingFood){
                 await API.put(`/foods/${editingFood._id}`, form);
-                alert("Food updated!");
-                setEditingFood(null);
+                toast.success("Food updated!");
             } else{
                 await API.post("/foods", form);
-                alert("Food added!");
+                toast.success("Food added!");
             }
+            setEditingFood(null);
             //reset form
             setForm({
                 title:"",
@@ -49,34 +52,26 @@ function CookDashboard(){
             });
             fetchMyFoods();
         } catch(error){
-            alert.error(error.response.data.message);
+            toast.error("Action Failed");
         }
     };
 
     const fetchOrders = async ()=>{
         try{
             const res = await API.get("/orders/cook-orders");
-            const ordersData = res.data.orders;
-            const incoming = ordersData.filter(o=> o.status === "pending");
-            const active = ordersData.filter(o=>o.status === "accepted" || o.status === "preparing");
-            const completed = ordersData.filter(o=>o.status === "delivered");
-            const cancelled = ordersData.filter(o=>o.status === "cancelled");
+            const data = res.data.orders;
+            setOrders(data);
             
-            setIncomingOrders(incoming);
-            setActiveOrders(active);
-            setCompleteOrders(completed);
-            setCancelledOrders(cancelled);
+            const valid = data.filter(o=>o.status!=="cancelled");
+            
 
             //calculate stats
-            const validOrders = ordersData.filter(o=>o.status !== "cancelled");
-
-            const totalOrders = validOrders.length;
-            const pendingOrders = incomingOrders.length;
-            const revenue = validOrders.reduce((sum, o)=> sum+o.totalPrice, 0);
-
-            setStats({totalOrders, pendingOrders, revenue});
+            setStats({totalOrders: valid.length,
+                 pendingOrders: valid.filter(o=>o.status === "pending").length,
+                  revenue: valid.reduce((sum, o)=> sum+o.totalPrice, 0)
+            });
         } catch (error){
-            console.error(error);
+            toast.error(error.response?.data?.message);
         }
     };
 
@@ -90,7 +85,24 @@ function CookDashboard(){
             await API.put(`/orders/${orderId}/status`, {status});
             fetchOrders(); //refresh
         } catch(error){
-            alert(error.response.data.message);
+            toast.error(error.response?.data?.message);
+        }
+    }
+
+    const getTabOrders =()=>{
+        switch(activeTab){
+            case "incoming":
+                return filteredOrders.filter(o=> o.status === "pending");
+            case "active":
+                return filteredOrders.filter(o=>
+                    ["accepted", "preparing", "delivering"].includes(o.status)
+                );
+            case "completed":
+                return filteredOrders.filter(o=> o.status === "delivered");
+            case "cancelled":
+                return filteredOrders.filter(o=> o.status === "cancelled");
+            default:
+                return [];
         }
     }
 
@@ -100,29 +112,35 @@ function CookDashboard(){
             const res = await API.get("/foods/my-foods");
             setMyFoods(res.data.foods);
         } catch(error) {
-            console.error(error);
+            toast.error(error.response?.data?.message);
         }
     }
 
+    const filteredOrders =orders.filter((o)=>
+    o.foodItem?.title.toLowerCase().includes(search.toLowerCase()));
+
     const handleDelete = async(id)=>{
+        if(!window.confirm("Are you sure you want to delete this food?")) return;
         try{
             await API.delete(`/foods/${id}`);
             fetchMyFoods();
         } catch(error) {
-            alert(error.response.data.message);
+            toast.error(error.response?.data?.message);
         }
     }
-    const toggleAvalability = async(food) =>{
-        try{
+    const toggleAvailability = async (food) => {
+        try {
             await API.put(`/foods/${food._id}`, {
-                available: !food.available
+            available: !food.available
             });
+
             fetchMyFoods();
 
-        } catch(error) {
-            alert(error.response.data.message);
+        } catch {
+            toast.error("Failed to update availability");
         }
     };
+
     const startEdit = (food) =>{
         setEditingFood(food);
         setForm({
@@ -137,111 +155,146 @@ function CookDashboard(){
             case "pending":
                 return { label: "Accept", next: "accepted"};
             case "accepted" :
-                return {label: "Mark as preparing", next:"preparing"};
+                return {label: "Preparing", next:"preparing"};
             case "preparing":
-                return {label:"Mark as Delivered", next: "delivered"};
+                return {label:"Out for Delivery", next: "delivering"};
+            case "delivering":
+                return {label: "Mark Delivered", next: "delivered"};
             default:
                 return null;
         }
     };
-    const OrderCard = ({ order }) =>{
-        const action = getNextStatus(order.status);
-        const isCompleted = order.status === "delivered";
-        return (
-            <div style={{border:"1px solid gray", margin: "10px"}}>
-                <p>Fodd: {order.foodItem?.title}</p>
-                <p>Student: {order.student?.name}</p>
-                <p>Quantity: {order.quantity}</p>
-                <p>Status: {order.status}</p>
-                {action && !isCompleted && (
-                    <button onClick={()=> updateStatus(order._id, action.next)}>
-                        {action.label}
-                    </button>
-                )}
-                {isCompleted && <p>Order Completed</p>}
+
+
+ return (
+        <div className="cook-container">
+            <div className="navbar">
+                <img src={logo} alt="Logo" className="logo" />
+                <input 
+                    className="search-bar" 
+                    placeholder="Search orders..." 
+                    value={search} 
+                    onChange={(e) => setSearch(e.target.value)} 
+                />
+                <button onClick={logout} className="logout-btn">Logout</button>
             </div>
-        )
-    }
 
-    const renderOrders = ()=>{
-        
-        let selectedOrders = [];
-        switch(activeTab){
-            case "incoming":
-                selectedOrders = incomingOrders;
-                break;
-            case "active":
-                selectedOrders = activeOrders;
-                break;
-            case "completed":
-                selectedOrders = completeOrders;
-                break;
-            case "cancelled":
-                selectedOrders = cancelledOrders;
-                break;
-            default:
-                selectedOrders = [];
-        }
-        if(selectedOrders.length === 0) {
-            return <p>No orders</p>
-        }
-        return selectedOrders.map((order)=>(
-            <OrderCard key={order._id} order={order}/>
-        ));
-    };
+            <h2 className="greeting">
+                Hi, {user?.user?.name} 👋
+            </h2>
 
-
-    return(
-        <div>
-            <h2>Cook Dashboard</h2>
-            <h3>Dashboard Stats</h3>
-            <p>Total Orders: {stats.totalOrders}</p>
-            <p>Pending Orders: {stats.pendingOrders}</p>
-            <p>Revenue: ₹{stats.revenue}</p>
-
-            <h3>{editingFood ? "Edit Food" : "Add Food"}</h3>
-            {editingFood && (
-                <p style={{color: "blue" }}>
-                    Currently Editing: <strong>{editingFood.title}</strong>
-                </p>
-            )}
-
-            <form onSubmit={handleSubmit}>
-                {/* react controls form state */}
-                <input name="title" placeholder="Name of food" value={form.title} onChange={handleChange}/> 
-                <input name="description" placeholder="Add description of food" value={form.description} onChange={handleChange}/>
-                <input name="price" placeholder="Price of food" value={form.price} onChange={handleChange}/>
-                <select name="isVeg" onChange={handleChange}>
-                    <option value="true">Veg</option>
-                    <option value= "false">Non-Veg</option>
-                </select>
-                <button type="submit">{editingFood ? "Update Food" : "Add Food"}</button>
-            </form>
-
-            <div  style={{ margin: "20px 0"}}>
-                <button style={{fontWeight:activeTab === "incoming" ? "bold":"normal"}}
-                onClick={()=>setActiveTab("incoming")}>Incoming</button>
-                <button style={{fontWeight: activeTab ==="active" ? "bold": "normal"}}
-                onClick={()=> setActiveTab("active")}>Active</button>
-                <button style={{ fontWeight: activeTab === "completed" ? "bold": "normal"}}
-                onClick={()=> setActiveTab("completed")}>Completed</button>
-                <button style={{fontWeight: activeTab === "cancelled" ? "bold": "normal"}}
-                onClick={()=> setActiveTab("cancelled")}>Cancelled</button>
-            </div>
-            <h3>{activeTab.toUpperCase()} ORDERS</h3>
-            {renderOrders()}
-
-            <h3>My Foods</h3>
-            {myFoods.map((food)=>(
-                <div key={food._id} style={{ border: "1px solid green", margin:"10px"}}>
-                    <p>{food.title}</p>
-                    <p>₹{food.price}</p>
-                    <p>{food.available ? "Available" : "Unavailable"}</p>
-                    <button onClick={()=> startEdit(food)}>Edit</button>
-                    <button onClick={()=> handleDelete(food._id)}>Delete</button>
-                    <button onClick={()=> toggleAvalability(food)}>Toggle Availability</button>
+            {/* stats */}
+            <div className="stats">
+                <div className="stat-card">
+                    <p>Total Orders</p>
+                    <h2>{stats.totalOrders}</h2>
                 </div>
-            ))}
+                <div className="stat-card">
+                    <p>Pending</p>
+                    <h2>{stats.pendingOrders}</h2>
+                </div>
+                <div className="stat-card">
+                    <p>Revenue</p>
+                    <h2>₹{stats.revenue}</h2>
+                </div>
+            </div>
+
+            {/* tabs */}
+            <div className="tabs">
+                {["incoming", "active", "completed", "cancelled", "foods"].map(tab => (
+                    <button 
+                        key={tab} 
+                        className={`tab ${activeTab === tab ? "active" : ""}`} 
+                        onClick={() => setActiveTab(tab)}
+                    >
+                        {tab.charAt(0).toUpperCase() + tab.slice(1)}
+                    </button>
+                ))}
+            </div>
+
+            {/* content */}
+            <div className="content">
+                {activeTab !== "foods" ? (
+                    getTabOrders().map(order => {
+                        const action = getNextStatus(order.status);
+                        return (
+                            <div className="order-card" key={order._id}>
+                                <div className="order-info">
+                                    <h3>{order.foodItem?.title}</h3>
+                                    <p>Student: {order.student?.name}</p>
+                                    <p>Qty: {order.quantity}</p>
+                                </div>
+                                {action && (
+                                    <button className="primary-btn" onClick={() => updateStatus(order._id, action.next)}>
+                                        {action.label}
+                                    </button>
+                                )}
+                            </div>
+                        );
+                    })
+                ) : (
+                    <div className="foods-section">
+                        <div className="food-form-card">
+                            <h3>{editingFood ? "Edit Food" : "Add New Dish"}</h3>
+                            {editingFood && (
+                                <p className="editing-text">
+                                    Editing: <strong>{editingFood.title}</strong>
+                                </p>
+                            )}
+                            <form onSubmit={handleSubmit}>
+                                <input
+                                    name="title"
+                                    value={form.title}
+                                    onChange={handleChange}
+                                    placeholder="Dish name" required
+                                />
+                                <input
+                                    name="description"
+                                    value={form.description}
+                                    onChange={handleChange}
+                                    placeholder="Short description" required
+                                />
+                                <input
+                                    name="price"
+                                    value={form.price}
+                                    onChange={handleChange}
+                                    placeholder="Price" required
+                                />
+                                <select name="isVeg" value={form.isVeg} onChange={handleChange}>
+                                    <option value="true">Veg</option>
+                                    <option value="false">Non-Veg</option>
+                                </select>
+
+                                <button className="primary-btn" type="submit">
+                                    {editingFood ? "Update Food" : "Add Food"}
+                                </button>
+                                {editingFood && (
+                                    <button type="button" className="secondary-btn" onClick={() => setEditingFood(null)}>
+                                        Cancel
+                                    </button>
+                                )}
+                            </form>
+                        </div>
+
+                        <div className="food-grid">
+                            {myFoods.map(food => (
+                                <div className="food-card" key={food._id}>
+                                    <h4>{food.title}</h4>
+                                    <p className="price-tag">₹{food.price}</p>
+                                    <p className={food.available ? "available" : "not-available"}>
+                                        {food.available ? "● Available" : "● Unavailable"}
+                                    </p>
+                                    <div className="food-actions">
+                                        <button className="edit-btn" onClick={() => startEdit(food)}>Edit</button>
+                                        <button className="delete-btn" onClick={() => handleDelete(food._id)}>Delete</button>
+                                        <button className="toggle-btn" onClick={() => toggleAvailability(food)}>Toggle</button>
+                                    </div>
+                                </div>
+                            ))}
+                        </div>
+                    </div>
+                )}
+            </div>
         </div>
     );
 }
